Parcel init and code refactor
